# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ma premi\u00e8re API",
        "description": "Une API permettant l'acc\u00e8s aux donn\u00e9es de la base via une approche RESTful",
        "termsOfService": "",
        "contact": {
            "name": "Guiks",
            "email": "guillaume.delapre@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "myapi.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/seance": {
            "get": {
                "tags": [
                    "seance"
                ],
                "summary": "Display a listing of seances.",
                "operationId": "indexSeance",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Seance"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "seance"
                ],
                "summary": "Create a seance",
                "description": "Use this method to create a seance",
                "operationId": "storeSeance",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_film",
                        "in": "formData",
                        "description": "Movie ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id_salle",
                        "in": "formData",
                        "description": "Room ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id_personne_ouvreur",
                        "in": "formData",
                        "description": "Room ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id_personne_technicien",
                        "in": "formData",
                        "description": "Technician guy ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id_personne_menage",
                        "in": "formData",
                        "description": "Cleaning guy ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "debut_seance",
                        "in": "formData",
                        "description": "Starting time",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "fin_seance",
                        "in": "formData",
                        "description": "Ending time",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Seance created",
                        "schema": {
                            "$ref": "#/definitions/Seance"
                        }
                    },
                    "422": {
                        "description": "Missing or incorrect fields"
                    }
                }
            }
        },
        "/seance/{id_seance}": {
            "get": {
                "tags": [
                    "seance"
                ],
                "summary": "Display a single seance",
                "description": "Use this method to return a single seance attributes based on its id.",
                "operationId": "showSeance",
                "parameters": [
                    {
                        "name": "id_seance",
                        "in": "path",
                        "description": "Seance ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Seance"
                        }
                    },
                    "404": {
                        "description": "Seance not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "seance"
                ],
                "summary": "Update a seance",
                "description": "Use this method to update the attributes of a seance based on its id.",
                "operationId": "updateSeance",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_seance",
                        "in": "path",
                        "description": "Seance ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id_film",
                        "in": "formData",
                        "description": "Movie ID",
                        "type": "integer"
                    },
                    {
                        "name": "id_salle",
                        "in": "formData",
                        "description": "Room ID",
                        "type": "integer"
                    },
                    {
                        "name": "id_personne_ouvreur",
                        "in": "formData",
                        "description": "Opener ID",
                        "type": "integer"
                    },
                    {
                        "name": "id_personne_technicien",
                        "in": "formData",
                        "description": "Technician ID",
                        "type": "integer"
                    },
                    {
                        "name": "id_personne_menage",
                        "in": "formData",
                        "description": "Cleaning ID",
                        "type": "integer"
                    },
                    {
                        "name": "debut_seance",
                        "in": "formData",
                        "description": "Starting time",
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "fin_seance",
                        "in": "formData",
                        "description": "Ending time",
                        "type": "string",
                        "maximum": "255"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seance updated",
                        "schema": {
                            "$ref": "#/definitions/Seance"
                        }
                    },
                    "422": {
                        "description": "Missing or incorrect fields"
                    }
                }
            },
            "delete": {
                "tags": [
                    "seance"
                ],
                "summary": "Delete a seance",
                "description": "Use this method to delete a seance based on its id.",
                "operationId": "destroySeance",
                "parameters": [
                    {
                        "name": "id_seance",
                        "in": "path",
                        "description": "Seance ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seance deleted"
                    },
                    "404": {
                        "description": "Seance not found"
                    }
                }
            }
        },
        "/seance/film/{id_film}": {
            "get": {
                "tags": [
                    "seance"
                ],
                "summary": "Display next seances by film id",
                "description": "Use this method to return a listing of next seances based on film id and dates.",
                "operationId": "getSeancesByIdFilm",
                "parameters": [
                    {
                        "name": "id_film",
                        "in": "path",
                        "description": "Film ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "date_debut",
                        "in": "query",
                        "description": "Starting time",
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "description": "Ending time",
                        "type": "string",
                        "format": "date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Seance"
                            }
                        }
                    },
                    "204": {
                        "description": "Successful operation but there isn't seance with this film or dates",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Seance"
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Seance": {
            "required": [
                "id_film",
                "id_salle",
                "id_personne_ouvreur",
                "id_personne_technicien",
                "id_personne_menage",
                "debut_seance",
                "fin_seance"
            ],
            "properties": {
                "id": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "id_film": {
                    "type": "number",
                    "format": "int64",
                    "default": 112
                },
                "id_salle": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "id_personne_ouvreur": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "id_personne_technicien": {
                    "type": "number",
                    "format": "int64",
                    "default": 2
                },
                "id_personne_menage": {
                    "type": "number",
                    "format": "int64",
                    "default": 3
                },
                "debut_seance": {
                    "type": "date",
                    "format": "date",
                    "default": "2016-04-04 10:25:00"
                },
                "fin_seance": {
                    "type": "date",
                    "format": "date",
                    "default": "2016-04-04 12:05:00"
                }
            },
            "xml": {
                "name": "Seance"
            }
        }
    }
}